% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\docType{data}
\name{ACE}
\alias{ACE}
\title{Function: printGeneInfo()}
\format{
An object of class \code{character} of length 1.
}
\usage{
ACE
}
\value{
Returns list of different genes and their descriptions.
}
\description{
Purpose: prints information on each gene that is being used in the
TGS calculation (name of gene, function of gene, variants and frequencies)
}
\keyword{datasets}
