% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalGeneticScore.R
\name{athleteProfile}
\alias{athleteProfile}
\title{Function: athleteProfile()}
\usage{
athleteProfile()
}
\value{
Returns a series of print functions to convey a detailed report on an
athlete's polygenic profile and its implications on training recommendations
and athletic performance.
}
\description{
Purpose: To generate a detailed report on an athlete's polygenic profile.
This function uses the helper function genotypeProb() and TGS calculations
to provide insights on an athlete's strengths, weakness' as well as consequential
effects on sports performances and individualized training recommendations.
}
