% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalGeneticScore.R
\name{calculateTGS}
\alias{calculateTGS}
\title{Helper function: calculateTGS()}
\usage{
calculateTGS()
}
\value{
Returns the total genotype score (TGS) of one polygenic profile.
}
\description{
Purpose: Total genotype score (TGS) reflects the additive effect of genotypes
on predicting a complex trait such as athletic performance. It is an algorithm
introduced by Williams and Folland and is also known as a genetic
predisposition score using synthetic data of 23 genetic markers.
}
\references{
Williams, A. G., & Folland, J. P. (2008). Similarity of polygenic profiles
limits the potential for elite human physical performance. The Journal of
physiology, 586(1), 113–121. https://doi.org/10.1113/jphysiol.2007.141887
}
