% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Sample Data of 23 Gene Polymorphisms and Cardiovascular Fitness}
\format{
## `dataset`
A data frame with 101 rows and 25 columns:
\describe{
  \item{ACE}{ACE}
  \item{ACTN3}{ACTN3}
  \item{ADRA2A}{ADRA2A}
  \item{ADRB2}{ADRB2}
  \item{AMPD1}{AMPD1}
  \item{APOE}{APOE}
  \item{ATP1A2 (exon 1)}{ATP1A2 (exon 1)}

  \item{Score}{Score}
  ...
}
}
\source{
<https://physoc.onlinelibrary.wiley.com/doi/full/10.1113/jphysiol.2007.141887>
}
\usage{
dataset
}
\description{
A fake data set generated from researched probabilities of each gene
polymorphism.
}
\keyword{datasets}
