% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{geneInfo}
\alias{geneInfo}
\title{Function: geneInfo()}
\usage{
geneInfo()
}
\value{
Returns list of different genes and their descriptions.
}
\description{
Purpose: prints information on each gene that is being used in the
TGS calculation (name of gene, function of gene, variants and frequencies).
}
\details{
Outputs the references used to gather information on the 23 gene polymorphisms.

# 'data.frame':    96 obs. of  8 variables:
}
