% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_overview.R
\name{geneOverview}
\alias{geneOverview}
\title{This function is used to provide an overview of a user's genetic profile.}
\usage{
geneOverview(file)
}
\arguments{
\item{file}{containing individual data set}
}
\value{
scatter plot with regression
}
\description{
The function categorizes each gene's variant into three buckets:
unfavorable, neutral and optimal for athletic performance.
}
\details{
This function returns three data frames summarizing the genes and their functions.
}
