% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalGeneticScore.R
\name{generateBoxPlot}
\alias{generateBoxPlot}
\title{Plot function: generateBoxPlot(n)}
\usage{
generateBoxPlot(n = 100)
}
\arguments{
\item{n}{for the number of polygenic profiles to be generated and their
respective calculated total genotype scores to be included in the dataset.}
}
\value{
Returns a frequency plot of n individuals and their respective TGS.
}
\description{
Purpose: To generate a box plot of n polygenic total genotype scores to
be used for analysis and plotting purposes.
}
