% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generateDataset}
\alias{generateDataset}
\title{generateDataset}
\usage{
generateDataset(entries)
}
\arguments{
\item{entries}{number of individuals to create data for}
}
\value{
a dataset in the form of a data frame with entries number of individuals
}
\description{
This function is used to generate and save a dataset of entries
containing 23 gene polymorphisms associatied with athletic endurance
performance ability as well as cardiovascular fitness level. Each gene and
their variants has a researched frequency in large populations. This dataset
uses these probabilities to generate a contribed dataset suitable for the
purposes of this R package. The cardiovascular fitness of each entry is based
on the genetic score.
}
\examples{
\dontrun{
generateDataset(100)
}

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language.
Wadsworth & Brooks/Cole.
}
