% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{genotypeHistogram}
\alias{genotypeHistogram}
\title{This function is used to visualize the frequencies of each genotype of the
the 23 genes provided in the data set.}
\usage{
genotypeHistogram(data)
}
\arguments{
\item{data}{containing data set to be plotted}
}
\value{
frequency plot
}
\description{
23 graphs will be output. Each graph representing a gene, each bar representing
a variant, and the y-axis representing the frequency.
}
