% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalGeneticScore.R
\name{genotypeProb}
\alias{genotypeProb}
\title{Purpose: Use contribed datasets of researched-backed polygenic information
to analyze and visualize predicted athletic endurance performances of
individuals based on their polygenic athlete profile and total genotype score.
Author: Helena Jovic
Date: November 11, 2022
Version: 1.0
Bugs and Issues:}
\usage{
genotypeProb()
}
\value{
a dataframe containing a column of genes associated with athletic
ability and a corresponding column of scores paired with each gene for one
individual.
}
\description{
Before each session run this in console
library(devtools)
library(roxygen2)
library(testthat)
library(knitr)
}
\details{
Helper function: genotypeProb()
A function that compiles genetic information of 23 genes associated with
endurance athletic ability. The function uses researched statistical data
on the frequencies of the genotypes of each gene to generate a polygenic
profile to predict athletic endurance performance. Each genotype is paired
with a respective genetic score, revealing how favourable (0 = unfavourable,
1 = neutral, 2 = favourable) each genotype is for athletic endurance ability.
}
\references{
Kambouris, M., Ntalouka, F., Ziogas, G., & Maffulli, N. (2012).
Predictive Genomics DNA Profiling for Athletic Performance. Recent patents
on DNA & gene sequences, 6. doi:10.2174/187221512802717321

Petr M, Thiel D, Kateřina K, Brož P, Malý T, Zahálka F, Vostatková P,
Wilk M, Chycki J, Stastny P. Speed and power-related gene polymorphisms
associated with playing position in elite soccer players. Biol Sport.
2022 Mar;39(2):355-366. doi: 10.5114/biolsport.2022.105333.
Epub 2021 Apr 21. PMID: 35309536; PMCID: PMC8919892.
}
