% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{genotypeScatter}
\alias{genotypeScatter}
\title{genotypeScatter}
\usage{
genotypeScatter(dataset, gene)
}
\arguments{
\item{dataset}{containing data set to be plotted}

\item{gene}{written in the format dataset$gene, corresponding to a column}
}
\value{
scatter plot with regression. X-axis: genotype score on scale of 0 to
2, where 0 means unfavourable genotype association with athletic performance,
1 means neutral genotype association with athletic performance and 2 means
optimal genotype association with athletic performance. Y-axis: cardiovascular
fitness on a scale of 1 lowest to 6 max based on VO2 max performance. A
regression line is added to show trend in data given.
}
\description{
This function is used to visualize the relationship between genotype score of
a particular gene in an individual and their cardiovascular fitness level. A
regression line is added to show the trend in the data given in the scatter plot.
}
\examples{
# Example:
# First load the dataset included in the package
ds <- system.file("extdata", "fake_dataset.xlsx", package = "PerfOMICS")
dataset <- readxl::read_excel(ds)

# Generate scatterplot using dataset
genotypeScatter(dataset,dataset$ACE)

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language.
Wadsworth & Brooks/Cole.Venables, W. N. and Ripley. B. D. (2002)
Modern Applied Statistics with S. Springer.
}
