% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerspeed_analysis.R
\name{powerSpeed}
\alias{powerSpeed}
\title{This function is used to provide user's a power speed analysis on their genetic
profile.}
\usage{
powerSpeed(file)
}
\arguments{
\item{file}{containing individual data set}
}
\value{
scatter plot with regression
}
\description{
The function checks for matches between a user's genetic profile and the optimal
variants for genes associated with power and speed ability.
}
\details{
This function returns the associated genotypes, their effect on sports performance
and recommended training.
}
