% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPerfOMICS.R
\name{runPerfOMICS}
\alias{runPerfOMICS}
\title{Launch Shiny App for PerfOMICS}
\usage{
runPerfOMICS()
}
\value{
opens a Shiny page
}
\description{
A function that launches the Shiny app for PerfOMICS, the code for the
Shiny app haas been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
PerfOMICS::runPerfOMICS()
}

}
