% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{This function is used to visualize a scatter plot of a variable in the provided
data set and the cardiovascular fitness level of each individual in the data set.
A regression line is added to show the trend in the data given in the scatter plot.}
\usage{
scatterPlot(data)
}
\arguments{
\item{data}{containing data set to be plotted}
}
\value{
scatter plot with regression
}
\description{
X-axis: Athletic Performance Score is based on 23 gene polymorphisms associated with athletic performance.
Y-axis: Cardiovascular Fitness on a scale of 1 lowest to 6 max based on VO2 max performance.
}
\details{
This function returns a scatter plot and a regression line.
}
\examples{
scatterPlot(dataset)
}
