% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo_plot.R
\name{scoreHistogram}
\alias{scoreHistogram}
\title{This function is used to visualize a histogram of the genetic scores
of each individual in the data set and a regression line.}
\usage{
scoreHistogram(data)
}
\arguments{
\item{data}{object containing data set to be plotted}
}
\value{
a histogram to visualize gene scores. X-axis is the athletic performance
score based on 23 gene polymorphisms associated with athletic performance.
Y-axis is the density of the gene scores. The graph outputs a density line.
}
\description{
This function is used to visualize a histogram of the genetic scores
of each individual in the data set and a regression line.
}
\examples{
# Example:

# First load the dataset included in the package
ds <- system.file("extdata", "fake_dataset.xlsx", package = "PerfOMICS")
data <- readxl::read_excel(ds)

# Generate histogram using dataset
scoreHistogram(data)

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language.
Wadsworth & Brooks/Cole.Venables, W. N. and Ripley. B. D. (2002)
Modern Applied Statistics with S. Springer.
}
