% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescore_analysis.R
\name{userScore}
\alias{userScore}
\title{This function is used to calculate a genetic score on a scale of 0 to 100
of an individual genetic profile of 23 gene polymorphisms.}
\usage{
userScore(file)
}
\arguments{
\item{file}{containing individual data set}
}
\value{
scatter plot with regression
}
\description{
The function uses an algorithm proposed by Williams and Folland in 2008,
which is based on researched associations between certain variants and
athletic endurance performance.
}
\details{
Understanding your score: a score of 0, means every gene in your genetic
profile is unfavourable for athletic endurance performance, whereas a score
of 100 means every gene in your genetic profile is optimal for athletic
endurance performance.
}
\references{
Kambouris, M., Ntalouka, F., Ziogas, G., & Maffulli, N. (2012).
Predictive Genomics DNA Profiling for Athletic Performance. Recent patents
on DNA & gene sequences, 6. doi:10.2174/187221512802717321

Petr M, Thiel D, Kateřina K, Brož P, Malý T, Zahálka F, Vostatková P,
Wilk M, Chycki J, Stastny P. Speed and power-related gene polymorphisms
associated with playing position in elite soccer players. Biol Sport.
2022 Mar;39(2):355-366. doi: 10.5114/biolsport.2022.105333.
Epub 2021 Apr 21. PMID: 35309536; PMCID: PMC8919892.
}
