% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalGeneticScore.R
\name{athleteProfile}
\alias{athleteProfile}
\title{Function: athleteProfile()}
\usage{
athleteProfile()
}
\value{
Returns a series of print functions to convey a detailed report on an
athlete's polygenic profile and its implications on training recommendations
and athletic performance.
}
\description{
Purpose: To generate a detailed report on an athlete's polygenic profile.
This function uses the helper function genotypeProb() and the total genotype
score calculation to provide insights on an athlete's strenghts, weakness'
as well as any meaningful effects on sports performances and training
recommendations.
}
