% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/dataset.R, R/generatedataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Sample Data of 23 Gene Polymorphisms and Cardiovascular Fitness}
\format{
## `dataset`
A data frame with 101 rows and 25 columns:
\describe{
  \item{ACE}{ACE}
  \item{ACTN3}{ACTN3}
  \item{ADRA2A}{ADRA2A}
  \item{ADRB2}{ADRB2}
  \item{AMPD1}{AMPD1}
  \item{APOE}{APOE}
  \item{ATP1A2 (exon 1)}{ATP1A2 (exon 1)}

  \item{Score}{Score}
  ...
}
}
\source{
<https://physoc.onlinelibrary.wiley.com/doi/full/10.1113/jphysiol.2007.141887>
}
\usage{
dataset

dataset(entries)

dataset(entries)
}
\arguments{
\item{entries}{number of individuals to create data for}
}
\value{
a data frame object with entries number of individuals

a saved dataset as an RDA file with entries number of individuals
}
\description{
A fake data set generated from researched probabilities of each gene
polymorphism.

Each gene and their variants has a researched frequency in large populations.
This dataset uses these probabilities to generate a contribed dataset suitable
for the purposes of this R package. The cardiovascular fitness of each entry
is based on the genetic score.
#'

Each gene and their variants has a researched frequency in large populations.
This dataset uses these probabilities to generate a contribed dataset suitable
for the purposes of this R package. The cardiovascular fitness of each entry
is based on the genetic score.
#'
}
\examples{
entries <- 100
dataset(entries)
head(ds)
}
\keyword{datasets}
