% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_overview.R
\name{geneOverview}
\alias{geneOverview}
\title{geneOverview}
\usage{
geneOverview(data)
}
\arguments{
\item{data}{containing individual data set of 23 gene polymorphisms}
}
\value{
this function returns three saved data frames summarizing the genes
and their functions
}
\description{
This function is used to provide an overview of a user's genetic profile by
categorizing each gene's variant into three buckets: unfavorable, neutral
and optimal for athletic endurance performance.
}
\examples{
# Example:

# First load the individual dataset included in the package
ds <- system.file("extdata", "Ind2.xlsx", package = "PerfOMICS")
data <- readxl::read_excel(ds)

# Generate summary using dataset
geneOverview(data)

}
