% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalGeneticScore.R
\name{generateDataSet}
\alias{generateDataSet}
\title{Function: generateDataSet(n)}
\usage{
generateDataSet(n)
}
\arguments{
\item{n}{for the number of polygenic profiles to be generated and their
respective calculated total genotype scores to be included in the dataset.}
}
\value{
Returns a dataframe of n individuals and their respective TGS.
}
\description{
Purpose: To generate a large dataset in the form of a data fram of n
polygenic total genotype scores to be used for analysis and (potential)
plotting purposes.
}
