% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerspeed_analysis.R
\name{powerSpeed}
\alias{powerSpeed}
\title{This function is used to provide user's a power speed analysis on their genetic
profile by checking for matches between a user's genetic profile and the optimal
variants for genes associated with power and speed ability.}
\usage{
powerSpeed(data)
}
\arguments{
\item{data}{containing individual data set}
}
\value{
the associated genotypes in an individual profile, each genotype's
effect on sports performance and the corresponding recommended training.
}
\description{
This function is used to provide user's a power speed analysis on their genetic
profile by checking for matches between a user's genetic profile and the optimal
variants for genes associated with power and speed ability.
}
\examples{
# Example:

# First load the individual dataset included in the package
ds <- system.file("extdata", "Ind2.xlsx", package = "PerfOMICS")
data <- readxl::read_excel(ds)

# Generate histogram using dataset
powerSpeed(data)

}
