% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{scatterPlot}
\usage{
scatterPlot(data)
}
\arguments{
\item{data}{containing data set to be plotted}
}
\value{
scatter plot with regression. X-axis: athletic performance score
based on 23 gene polymorphisms associated with athletic performance.
Y-axis: cardiovascular fitness on a scale of 1 lowest to 6 max based on
VO2 max performance. A regression line is added to show trend in data given.
}
\description{
This function is used to visualize a scatter plot of a variable in the provided
data set and the cardiovascular fitness level of each individual in the data set.
A regression line is added to show the trend in the data given in the scatter plot.
}
\examples{
# Example:
# First load the dataset included in the package
ds <- system.file("extdata", "fake_dataset.xlsx", package = "PerfOMICS")
data <- readxl::read_excel(ds)

# Generate scatterplot using dataset
scatterPlot(data)

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language.
Wadsworth & Brooks/Cole.Venables, W. N. and Ripley. B. D. (2002)
Modern Applied Statistics with S. Springer.
}
